backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.1
langchain==0.3.23
langchain-openai==0.3.12
langchain-community==0.3.21
pydantic==2.11.3
pydantic-core==2.33.1
python-dotenv==1.1.0
pytest==7.4.3
httpx==0.25.2
pgvector==0.2.3
psycopg2-binary==2.9.9
pinecone-client==2.2.4
sentence-transformers==2.2.2
pymupdf==1.23.8
tqdm==4.66.1
tiktoken==0.9.0
requests==2.31.0
openai==1.74.0